requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.0
docxtpl==0.16.8
python-docx==1.1.0
